Rogerio_06_07_17_00

+++ src/main/java/br/com/tisyserp/model/estoque/Estoque.java
@@ -11,7 +11,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
@@ -21,7 +20,7 @@ import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 
 public class Estoque extends PanacheEntityBase {
     
-    public static  BigDecimal serialVersionUID = 1L;
+    public static  Long serialVersionUID = 1L;
     @Id
     @Basic(optional = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
